requirements.txt
numpy
Pillow
matplotlib
opencv-python
ultralytics
tqdm
icrawler
streamlit
watchdog
